refactor: create a sidebar folder
